+++ lib/hanami/model/associations/belongs_to.rb
@@ -5,36 +5,36 @@ module Hanami
     module Associations
       # Many-To-One association
       #
-      # @since x.x.x
+      # @since 1.1.0
       # @api private
       class BelongsTo
-        # @since x.x.x
+        # @since 1.1.0
         # @api private
         def self.schema_type(entity)
           Sql::Types::Schema::AssociationType.new(entity)
         end
 
-        # @since x.x.x
+        # @since 1.1.0
         # @api private
         attr_reader :repository
 
-        # @since x.x.x
+        # @since 1.1.0
         # @api private
         attr_reader :source
 
-        # @since x.x.x
+        # @since 1.1.0
         # @api private
         attr_reader :target
 
-        # @since x.x.x
+        # @since 1.1.0
         # @api private
         attr_reader :subject
 
-        # @since x.x.x
+        # @since 1.1.0
         # @api private
         attr_reader :scope
 
-        # @since x.x.x
+        # @since 1.1.0
         # @api private
         def initialize(repository, source, target, subject, scope = nil)
           @repository = repository
@@ -45,7 +45,7 @@ module Hanami
           freeze
         end
 
-        # @since x.x.x
+        # @since 1.1.0
         # @api private
         def one
           scope.one
@@ -53,25 +53,25 @@ module Hanami
 
         private
 
-        # @since x.x.x
+        # @since 1.1.0
         # @api private
         def container
           repository.container
         end
 
-        # @since x.x.x
+        # @since 1.1.0
         # @api private
         def primary_key
           association_keys.first
         end
 
-        # @since x.x.x
+        # @since 1.1.0
         # @api private
         def relation(name)
-          repository.relations[Hanami::Utils::String.new(name).pluralize]
+          repository.relations[Hanami::Utils::String.pluralize(name)]
         end
 
-        # @since x.x.x
+        # @since 1.1.0
         # @api private
         def foreign_key
           association_keys.last
@@ -79,18 +79,25 @@ module Hanami
 
         # Returns primary key and foreign key
         #
-        # @since x.x.x
+        # @since 1.1.0
         # @api private
         def association_keys
-          relation(source)
-            .associations[target]
+          association
             .__send__(:join_key_map, container.relations)
         end
 
-        # @since x.x.x
+        # Return the ROM::Associations for the source relation
+        #
+        # @since 1.1.9
+        # @api private
+        def association
+          relation(source).associations[target]
+        end
+
+        # @since 1.1.0
         # @api private
         def _build_scope
-          result = relation(target)
+          result = relation(association.target.to_sym)
           result = result.where(foreign_key => subject.fetch(primary_key)) unless subject.nil?
           result.as(Model::MappedRelation.mapper_name)
         end