+++ lib/hanami/model/associations/many_to_many.rb
@@ -1,3 +1,5 @@
+require "hanami/utils/hash"
+
 module Hanami
   module Model
     module Associations
@@ -5,7 +7,7 @@ module Hanami
       #
       # @since 0.7.0
       # @api private
-      class ManyToMany
+      class ManyToMany # rubocop:disable Metrics/ClassLength
         # @since 0.7.0
         # @api private
         def self.schema_type(entity)
@@ -13,27 +15,27 @@ module Hanami
           Types::Strict::Array.member(type)
         end
 
-        # @since x.x.x
+        # @since 1.1.0
         # @api private
         attr_reader :repository
 
-        # @since x.x.x
+        # @since 1.1.0
         # @api private
         attr_reader :source
 
-        # @since x.x.x
+        # @since 1.1.0
         # @api private
         attr_reader :target
 
-        # @since x.x.x
+        # @since 1.1.0
         # @api private
         attr_reader :subject
 
-        # @since x.x.x
+        # @since 1.1.0
         # @api private
         attr_reader :scope
 
-        # @since x.x.x
+        # @since 1.1.0
         # @api private
         attr_reader :through
 
@@ -67,50 +69,59 @@ module Hanami
           __new__(scope.where(condition))
         end
 
-        # @since x.x.x
-        # @api private
         # Return the association table object. Would need an aditional query to return the entity
-        def add(*data) # Can receive an array of hashes with pks.
+        #
+        # @since 1.1.0
+        # @api private
+        def add(*data)
           command(:create, relation(through), use: [:timestamps])
-            .call(associate(data.map(&:to_h)))
+            .call(associate(serialize(data)))
+        rescue => e
+          raise Hanami::Model::Error.for(e)
         end
 
-        # @since x.x.x
+        # @since 1.1.0
         # @api private
-        # disabled until I figure out a better way to do this
+        def delete
+          relation(through).where(source_foreign_key => subject.fetch(source_primary_key)).delete
+        end
+
+        # @since 1.1.0
+        # @api private
         # rubocop:disable Metrics/AbcSize
-        def remove(id)
+        def remove(target_id)
           association_record = relation(through)
-                               .where(target_foreign_key => id, source_foreign_key => subject.fetch(source_primary_key))
+                               .where(target_foreign_key => target_id, source_foreign_key => subject.fetch(source_primary_key))
                                .one
-          if association_record
-            ar_id = association_record.public_send relation(through).primary_key
-            command(:delete, relation(through), use: [:timestamps]).by_pk(ar_id).call
-          end
+
+          return if association_record.nil?
+
+          ar_id = association_record.public_send relation(through).primary_key
+          command(:delete, relation(through)).by_pk(ar_id).call
         end
         # rubocop:enable Metrics/AbcSize
 
         private
 
-        # @since x.x.x
+        # @since 1.1.0
         # @api private
         def container
           repository.container
         end
 
-        # @since x.x.x
+        # @since 1.1.0
         # @api private
         def relation(name)
           repository.relations[name]
         end
 
-        # @since x.x.x
+        # @since 1.1.0
         # @api private
         def command(target, relation, options = {})
           repository.command(target, relation, options)
         end
 
-        # @since x.x.x
+        # @since 1.1.0
         # @api private
         def associate(data)
           relation(target)
@@ -118,19 +129,19 @@ module Hanami
             .associate(container.relations, data, subject)
         end
 
-        # @since x.x.x
+        # @since 1.1.0
         # @api private
         def source_primary_key
           association_keys[0].first
         end
 
-        # @since x.x.x
+        # @since 1.1.0
         # @api private
         def source_foreign_key
           association_keys[0].last
         end
 
-        # @since x.x.x
+        # @since 1.1.0
         # @api private
         def association_keys
           relation(source)
@@ -138,23 +149,32 @@ module Hanami
             .__send__(:join_key_map, container.relations)
         end
 
-        # @since x.x.x
+        # @since 1.1.0
         # @api private
         def target_foreign_key
           association_keys[1].first
         end
 
-        # @since x.x.x
+        # @since 1.1.0
         # @api private
         def target_primary_key
           association_keys[1].last
         end
 
-        # @since x.x.x
+        # Return the ROM::Associations for the source relation
+        #
+        # @since 1.1.0
+        # @api private
+        def association
+          relation(source).associations[target]
+        end
+
+        # @since 1.1.0
+        #
         # @api private
         # rubocop:disable Metrics/AbcSize
         def _build_scope
-          result = relation(target).qualified
+          result = relation(association.target.to_sym).qualified
           unless subject.nil?
             result = result
                      .join(through, target_foreign_key => target_primary_key)
@@ -164,10 +184,18 @@ module Hanami
         end
         # rubocop:enable Metrics/AbcSize
 
-        # @since x.x.x
+        # @since 1.1.0
         # @api private
         def __new__(new_scope)
           self.class.new(repository, source, target, subject, new_scope)
+        end
+
+        # @since 1.1.0
+        # @api private
+        def serialize(data)
+          data.map do |d|
+            Utils::Hash.deep_serialize(d)
+          end
         end
       end
     end