+++ lib/hanami/model/sql/types.rb
@@ -70,7 +70,7 @@ module Hanami
 
             # NOTE: In the future rom-sql should be able to always return Ruby
             # types instead of Sequel types. When that will happen we can get
-            # rid of this logic in the block and to fallback to:
+            # rid of this logic in the block and fall back to:
             #
             #  MAPPING.fetch(unwrapped.pristine, attribute)
             MAPPING.fetch(unwrapped.pristine) do
@@ -82,11 +82,21 @@ module Hanami
             end
           end
 
+          # @since 1.0.4
+          # @api private
+          def self.pg_json_pristines
+            @pg_json_pristines ||= ::Hash.new do |hash, type|
+              hash[type] = if defined?(ROM::SQL::Types::PG)
+                             ROM::SQL::Types::PG.const_get(type).pristine
+                           end
+            end
+          end
+
           # @since 1.0.2
           # @api private
           def self.pg_json?(pristine)
-            (defined?(ROM::SQL::Types::PG::JSONB) && pristine == ROM::SQL::Types::PG::JSONB) ||
-              (defined?(ROM::SQL::Types::PG::JSON) && pristine == ROM::SQL::Types::PG::JSON)
+            pristine == pg_json_pristines['JSONB'.freeze] ||
+              pristine == pg_json_pristines['JSON'.freeze]
           end
 
           private_class_method :pg_json?