+++ spec/unit/hanami/model/version_spec.rb
@@ -1,5 +1,5 @@
 RSpec.describe "Hanami::Model::VERSION" do
-  it 'exposes version' do
-    expect(Hanami::Model::VERSION).to eq('1.0.2')
+  it "exposes version" do
+    expect(Hanami::Model::VERSION).to eq("1.1.0")
   end
 end